move entity state controller to its folder
